Rejoining BigInteger native operators into main file
