+++ src/Deveel.Math.Core/Deveel.Math.Core.csproj
@@ -39,7 +39,6 @@
     <Compile Include="..\Deveel.Math\Deveel.Math\BigInteger_Convertible.cs" Link="Math\BigInteger_Convertible.cs" />
     <Compile Include="..\Deveel.Math\Deveel.Math\BigInteger_Functions.cs" Link="Math\BigInteger_Functions.cs" />
     <Compile Include="..\Deveel.Math\Deveel.Math\BigMath.cs" Link="Math\BigMath.cs" />
-    <Compile Include="..\Deveel.Math\Deveel.Math\BitInteger_Operators.cs" Link="Math\BitInteger_Operators.cs" />
     <Compile Include="..\Deveel.Math\Deveel.Math\BitLevel.cs" Link="Math\BitLevel.cs" />
     <Compile Include="..\Deveel.Math\Deveel.Math\CharHelper.cs" Link="Math\CharHelper.cs" />
     <Compile Include="..\Deveel.Math\Deveel.Math\Conversion.cs" Link="Math\Conversion.cs" />
