+++ src/Deveel.Math/Deveel.Math.licenseheader
@@ -1,7 +1,7 @@
 extensions: designer.cs generated.cs
 extensions: .cs .cpp .h
 // 
-//  Copyright 2009-2014  Deveel
+//  Copyright 2009-2017  Deveel
 // 
 //    Licensed under the Apache License, Version 2.0 (the "License");
 //    you may not use this file except in compliance with the License.
