+++ src/Deveel.Math/Properties/AssemblyInfo.cs
@@ -1,5 +1,5 @@
 ﻿// 
-//  Copyright 2009-2014  Deveel
+//  Copyright 2009-2017  Deveel
 // 
 //    Licensed under the Apache License, Version 2.0 (the "License");
 //    you may not use this file except in compliance with the License.
@@ -12,6 +12,7 @@
 //    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 //    See the License for the specific language governing permissions and
 //    limitations under the License.
+
 using System.Reflection;
 using System.Resources;
 using System.Runtime.CompilerServices;
